server/requirements.txt
fastapi
uvicorn[standard]
openai
tiktoken
pywebview